refactor: move the LSP-type workspace helper to the LWC LSP (its only consumer)
