+++ packages/salesforcedx-lightning-lsp-common/src/shared.ts
@@ -6,8 +6,6 @@
  */
 
 import * as path from 'node:path';
-import { LspFileSystemAccessor } from './providers/lspFileSystemAccessor';
-import { readPackageJson } from './utils';
 
 const SFDX_PROJECT = 'sfdx-project.json';
 
@@ -25,110 +23,3 @@ export const isLWC = (type: WorkspaceType): boolean =>
   type === 'SFDX' || type === 'STANDARD_LWC' || type === 'CORE_ALL' || type === 'CORE_PARTIAL';
 
 export const getSfdxProjectFile = (root: string): string => path.join(root, SFDX_PROJECT);
-
-/**
- * @param root
- * @returns WorkspaceType for singular root
- */
-export const detectWorkspaceHelper = async (
-  root: string,
-  fileSystemAccessor: LspFileSystemAccessor
-): Promise<WorkspaceType> => {
-  try {
-    const sfdxProjectFile = getSfdxProjectFile(root);
-    const fileStat = await fileSystemAccessor.getFileStat(sfdxProjectFile);
-
-    if (fileStat?.type === 'file') {
-      return 'SFDX';
-    }
-  } catch {
-    // File doesn't exist, continue
-  }
-
-  try {
-    const fileStat = await fileSystemAccessor.getFileStat(`${path.join(root, 'workspace-user.xml')}`);
-    if (fileStat?.type === 'file') {
-      return 'CORE_ALL';
-    }
-  } catch {
-    // File doesn't exist, continue
-  }
-
-  try {
-    const parentWorkspaceUserUri = path.join(root, '..', 'workspace-user.xml');
-    const fileStat = await fileSystemAccessor.getFileStat(`${parentWorkspaceUserUri}`);
-    if (fileStat?.type === 'file') {
-      return 'CORE_PARTIAL';
-    }
-  } catch {
-    // File doesn't exist, continue
-  }
-
-  try {
-    const lwcConfigUri = path.join(root, 'lwc.config.json');
-    const fileStat = await fileSystemAccessor.getFileStat(`${lwcConfigUri}`);
-    if (fileStat?.type === 'file') {
-      return 'STANDARD_LWC';
-    }
-  } catch {
-    // File doesn't exist, continue
-  }
-
-  try {
-    const packageInfo = await readPackageJson(root, fileSystemAccessor);
-    if (!packageInfo) {
-      throw new Error('Package info not found');
-    }
-    const dependencies = Object.keys(packageInfo.dependencies ?? {});
-    const devDependencies = Object.keys(packageInfo.devDependencies ?? {});
-    const allDependencies: string[] = [...dependencies, ...devDependencies];
-    const hasLWCdependencies = allDependencies.some(key => key.startsWith('@lwc/') || key === 'lwc');
-
-    // any type of @lwc is a dependency
-    if (hasLWCdependencies) {
-      return 'STANDARD_LWC';
-    }
-
-    // has any type of lwc configuration
-    if (packageInfo.lwc) {
-      return 'STANDARD_LWC';
-    }
-
-    if (packageInfo.workspaces) {
-      return 'MONOREPO';
-    }
-
-    try {
-      const lernaJsonUri = path.join(root, 'lerna.json');
-      const fileStat = await fileSystemAccessor.getFileStat(`${lernaJsonUri}`);
-      if (fileStat?.type === 'file') {
-        return 'MONOREPO';
-      }
-    } catch {
-      // File doesn't exist, continue
-    }
-
-    return 'STANDARD';
-  } catch {
-    // Log error and fallback to setting workspace type to Unknown
-  }
-
-  return 'UNKNOWN';
-};
-
-/**
- * @param workspaceRoots
- * @param fileSystemAccessor
- * @returns WorkspaceType — multi-root only supported when all roots are CORE_PARTIAL
- */
-export const detectWorkspaceType = async (
-  workspaceRoots: string[],
-  fileSystemAccessor: LspFileSystemAccessor
-): Promise<WorkspaceType> => {
-  const types = await Promise.all(workspaceRoots.map(root => detectWorkspaceHelper(root, fileSystemAccessor)));
-  return types.length === 1
-    ? types[0]
-    : types.every(t => t === 'CORE_PARTIAL')
-      ? 'CORE_PARTIAL'
-      : 'UNKNOWN';
-};
